backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.24
albumentations @ git+https://github.com/albumentations-team/albumentations@b59030cb5a4d03ad57dca6ce4f985e0d9d8c0dd7
altair==5.4.1
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
category_encoders==2.7.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cryptography==45.0.7
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
Django==5.0.6
django-colorfield==0.11.0
django-cors-headers==4.7.0
django-otp==1.5.0
django-tailwind==3.8.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
dlib @ file:///C:/Users/aring/OneDrive/Desktop/proctor/proctor-be/img_inferences/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.111.1
fastapi-cli==0.0.11
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==25.1.21
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.8.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.23.5
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
imutils==0.5.4
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.3.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.7
langchain-core==1.0.0
langchain-google-genai==3.0.0
langcodes==3.5.0
langsmith==0.4.37
language_data==1.3.0
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
marisa-trie==1.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.7.5
mcp==1.14.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.6.0
motor==3.7.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.13
namex==0.0.8
narwhals==1.14.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.2
nh3==0.2.20
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
openai==1.35.15
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
patsy==1.0.1
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.4.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pkginfo==1.12.0
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
preshed==3.0.10
prometheus_client==0.21.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.5
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycaret==3.3.2
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.10.1
pydantic_core==2.18.4
pydeck==0.9.1
PyGithub==2.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.10.1
PyNaCl==1.6.0
pyod==2.0.3
pyparsing==3.1.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2022.1
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==2.0.13
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.15.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.3
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==2.5.1
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==3.0.2
starlette==0.37.2
statsmodels==0.14.4
streamlit==1.35.0
stringzilla==3.12.5
sympy==1.13.1
ta-lib @ file:///D:/ta_lib-0.6.3-cp310-cp310-win_amd64.whl#sha256=effb277b2470f44e41cd2ac350b2c73b790c6ca1f8f0cd65264f97358f1b8e24
tbats==1.1.3
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
timm==1.0.15
tinycss2==1.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.2
torch==2.1.2
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.39.3
tsdownsample==0.1.3
twine==6.0.1
typer==0.19.2
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
vboxapi==1.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
weasel==0.3.4
webcolors==24.8.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.6
whitenoise==6.6.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
yellowbrick==1.5
zipp==3.21.0
zstandard==0.25.0
